feat: dataclass-based tasks and task results
The `Task` and `TaskResult` classes are now `dataclass`es.

Dataclasses automatically implement dunder attributes and methods,
which simplifies writing and working with new tasks and task results.
Primary benefits include:

1. Automatic implementation of `__slots__`, `__init__`, and `__eq__`.
2. Automatic conversion of `TaskResult` to `dict` (using `dataclasses.asdict()`).
3. Declaration of attributes and their types in the class body instead of `__slots__`.

The follow convenience changes, unrelated to dataclasses, are also included:

**Task**

1. The new `_result_cls` class variable is used to
instaniate a `TaskResult` instance in the `Task` constructor.
Previously, `Task.result` was `Optional` was not set by default,
which meant having to override `gen_task_result`,
and manually set the attribute/check for it before using it.
2. `_EXCLUDE_FROM_MERGE` is now `_exclude_from_merge`.
3. The new attribute `raise_exceptions`, when set to `True`,
will cause `Task._perform_task` to raise any exception instead of
capturing it on the result.

**CliTask**

1. `CLITask` is now `CliTask`.
2. `COMMAND` is now `command`.
3. `DESCRIPTION` is now `description`.
4. The new `aliases` attribute allows defining CLI command aliases
(e.g. `["w", "wo"]` for the command "worms").
5. The new `_task_cls` class variable is used to by the default
`_perform_task` implementation to create an instance of a `Task`,
merge in the parsed command line arguments, and run the task.
This allows simply defining the CLI parameters and the task to run
instead of re-implementing `_perform_task` every time.
6. `_gen_task` allows overriding how `args` are merged with `_task_cls`
(e.g. to change the type of specific command line arguments).

**gen_cli_parser**

1. For nesting tasks under a subcommand, you can now provide subcommand aliases.
See the docstring of `gen_cli_parser` for examples.